codes/surgical-video-processing/requirements.txt
# Python dependencies for surgical video processing framework

# Core video processing
opencv-python>=4.8.0
numpy>=1.21.0
scikit-image>=0.19.0
Pillow>=9.0.0

# Configuration management
PyYAML>=6.0
pydantic>=1.10.0

# Data processing and analysis
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging and utilities
tqdm>=4.64.0
click>=8.0.0

# Optional: GPU acceleration (uncomment if using GPU)
# torch>=1.12.0
# torchvision>=0.13.0

# Optional: Advanced video analysis (uncomment if needed)
# moviepy>=1.0.3
# imageio>=2.20.0

# Development dependencies (install separately with pip install -r requirements-dev.txt)
# pytest>=7.0.0
# flake8>=5.0.0
# mypy>=0.991
# black>=22.0.0
# isort>=5.10.0
